+++ NEWS.md
@@ -1,3 +1,8 @@
+# incR 2.0.0.9000
+* New function added, incRscan_v2. This new function applies a fixed temperature threshold to assign on/off-bouts. incRscan_v2 may handle differences in data quality among days worse than the original incRscan function but it recovers on/off-bouts information for every day in the original data set (unlike incRscan which starts the analysis in the first day with complete data).
+* incRenv function re-coded to process data much faster
+* Small bug in incRscan fixed
+
 # incR 1.1.5.9000
 * incRscan code amended to handle missing data better
 * incRatt code amended to confirm adequate date format